requirements.txt
streamlit
watchdog
transformers
torch
torchvision
pillow